Add a mixin for text channel.
This is a prep for production rule (Forgot to include in the previous PR)
