+++ src/compile/mark/text.ts
@@ -1,10 +1,10 @@
-import {TEXT, X} from '../../channel';
-import {numberFormat, timeFormatExpression, getMarkConfig} from '../common';
+import {X} from '../../channel';
+import {getMarkConfig} from '../common';
 
 import * as mixins from './mixins';
 import {Config} from '../../config';
-import {ChannelDef, TextFieldDef, ValueDef, field, isFieldDef} from '../../fielddef';
-import {QUANTITATIVE, TEMPORAL} from '../../type';
+import {ChannelDef, isFieldDef} from '../../fielddef';
+import {QUANTITATIVE} from '../../type';
 import {UnitModel} from '../unit';
 import {VgValueRef} from '../../vega.schema';
 
@@ -23,7 +23,7 @@ export const text: MarkCompiler = {
     return {
       ...mixins.pointPosition('x', model, xDefault(config, textDef)),
       ...mixins.pointPosition('y', model, ref.midY(config)),
-      text: textRef(textDef, config),
+      ...mixins.text(model),
       ...mixins.color(model),
       ...mixins.nonPosition('opacity', model),
       ...mixins.nonPosition('size', model, {
@@ -42,36 +42,11 @@ function xDefault(config: Config, textDef: ChannelDef): VgValueRef {
   return {value: config.scale.textXRangeStep / 2};
 }
 
-function textRef(textDef: TextFieldDef | ValueDef<any>, config: Config): VgValueRef {
-  // text
-  if (textDef) {
-    if (isFieldDef(textDef)) {
-      if (QUANTITATIVE === textDef.type) {
-        // FIXME: what happens if we have bin?
-        const format = numberFormat(textDef, textDef.format, config, TEXT);
-        return {
-          signal: `format(${field(textDef, {datum: true})}, '${format}')`
-        };
-      } else if (TEMPORAL === textDef.type) {
-        return {
-          signal: timeFormatExpression(field(textDef, {datum: true}), textDef.timeUnit, textDef.format, config.text.shortTimeLabels, config.timeFormat)
-        };
-      } else {
-        return {field: textDef.field};
-      }
-    } else if (textDef.value) {
-      return {value: textDef.value};
-    }
-  }
-  return {value: config.text.text};
-}
-
 function align(encoding: Encoding, config: Config) {
   const alignConfig = getMarkConfig('align', 'text', config);
   if (alignConfig === undefined) {
     return channelHasField(encoding, X) ? 'center' : 'right';
   }
-
   // If there is a config, Vega-parser will process this already.
   return undefined;
 }
