+++ src/compile/mark/valueref.ts
@@ -4,11 +4,12 @@
 
 import {Channel, X, X2, Y, Y2} from '../../channel';
 import {Config} from '../../config';
-import {ChannelDef, FieldDef, FieldRefOption, field, isFieldDef} from '../../fielddef';
+import {ChannelDef, FieldDef, FieldRefOption, field, isFieldDef, TextFieldDef, ValueDef} from '../../fielddef';
 import {Scale, ScaleType, hasDiscreteDomain} from '../../scale';
 import {StackProperties} from '../../stack';
 import {contains} from '../../util';
 import {VgValueRef} from '../../vega.schema';
+import {numberFormat, timeFormatExpression} from '../common';
 
 // TODO: we need to find a way to refactor these so that scaleName is a part of scale
 // but that's complicated.  For now, this is a huge step moving forward.
@@ -124,6 +125,30 @@ export function midPoint(channel: Channel, channelDef: ChannelDef, scaleName: st
     }
   }
   return defaultRef;
+}
+
+export function text(textDef: TextFieldDef | ValueDef<any>, config: Config): VgValueRef {
+  // text
+  if (textDef) {
+    if (isFieldDef(textDef)) {
+      if (textDef.type === 'quantitative') {
+        // FIXME: what happens if we have bin?
+        const format = numberFormat(textDef, textDef.format, config, 'text');
+        return {
+          signal: `format(${field(textDef, {datum: true})}, '${format}')`
+        };
+      } else if (textDef.type === 'temporal') {
+        return {
+          signal: timeFormatExpression(field(textDef, {datum: true}), textDef.timeUnit, textDef.format, config.text.shortTimeLabels, config.timeFormat)
+        };
+      } else {
+        return {field: textDef.field};
+      }
+    } else if (textDef.value) {
+      return {value: textDef.value};
+    }
+  }
+  return {value: config.text.text};
 }
 
 export function midX(config: Config): VgValueRef {
